+++ CHANGES.md
@@ -37,6 +37,7 @@ New:
 - Added support for bash completion.
 - Added `video.add_text.native`.
 - Added `configure.bindir`
+- Added ffmpeg-based stream decoder, limited to mime type `application/ffmpeg` for now.
 
 Changed:
 
@@ -118,6 +119,7 @@ Fixed:
 - Fixed audio glitches when seeking within a MP3 file.
 - Fixed `insert_metadata` logic when insert new track and metadata (#903)
 - Fixed `replay-gain` script default location.
+- Fixed audio glitches at the end of crossfade transitions.
 
 1.3.7 (09-04-2019)
 =====